requirements.txt
Flask
gunicorn
sqlmodel
psycopg2-binary
requests
structlog

google-auth
pipecat-ai[openai,silero,deepgram,together,deepseek,grok,groq,anthropic,google,elevenlabs]
fastapi[standard]
uvicorn
python-dotenv
loguru
python-multipart
google-cloud-storage
twilio

black
flake8
pre-commit
